Move Theme events to core_event_dispatcher
