+++ modules/core_event_dispatcher/src/Event/Theme/PageBottomEvent.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Drupal\hook_event_dispatcher\Event\Page;
+namespace Drupal\core_event_dispatcher\Event\Theme;
 
 use Drupal\hook_event_dispatcher\Event\EventInterface;
 use Drupal\hook_event_dispatcher\HookEventDispatcherInterface;
 use Symfony\Component\EventDispatcher\Event;
 
 /**
- * Class PageTopEvent.
+ * Class PageBottomEvent.
  */
-class PageTopEvent extends Event implements EventInterface {
+class PageBottomEvent extends Event implements EventInterface {
 
   /**
    * The build array.
@@ -19,7 +19,7 @@ class PageTopEvent extends Event implements EventInterface {
   private $build;
 
   /**
-   * PageTopEvent constructor.
+   * PageBottomEvent constructor.
    *
    * @param array $build
    *   The build array.
@@ -42,7 +42,7 @@ class PageTopEvent extends Event implements EventInterface {
    * {@inheritdoc}
    */
   public function getDispatcherType(): string {
-    return HookEventDispatcherInterface::PAGE_TOP;
+    return HookEventDispatcherInterface::PAGE_BOTTOM;
   }
 
 }
