+++ modules/core_event_dispatcher/src/Event/Theme/ThemeSuggestionsAlterIdEvent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\hook_event_dispatcher\Event\Theme;
+namespace Drupal\core_event_dispatcher\Event\Theme;
 
 /**
  * Class ThemeSuggestionsAlterIdEvent.
