+++ modules/core_event_dispatcher/tests/src/Unit/Theme/ThemesInstalledEventTest.php
@@ -1,18 +1,16 @@
 <?php
 
-namespace Drupal\Tests\hook_event_dispatcher\Unit\Theme;
+namespace Drupal\Tests\core_event_dispatcher\Unit\Theme;
 
 use Drupal;
 use Drupal\Core\DependencyInjection\ContainerBuilder;
 use Drupal\hook_event_dispatcher\HookEventDispatcherInterface;
 use Drupal\Tests\hook_event_dispatcher\Unit\HookEventDispatcherManagerSpy;
 use Drupal\Tests\UnitTestCase;
-use function hook_event_dispatcher_themes_installed;
+use function core_event_dispatcher_themes_installed;
 
 /**
  * Class ThemesInstalledEventTest.
- *
- * @package Drupal\Tests\hook_event_dispatcher\Unit\Theme
  *
  * @group hook_event_dispatcher
  */
@@ -42,9 +40,9 @@ class ThemesInstalledEventTest extends UnitTestCase {
   public function testThemesInstalledEvent(): void {
     $themes = ['classy', 'bartik'];
 
-    hook_event_dispatcher_themes_installed($themes);
+    core_event_dispatcher_themes_installed($themes);
 
-    /* @var \Drupal\hook_event_dispatcher\Event\Theme\ThemesInstalledEvent $event */
+    /* @var \Drupal\core_event_dispatcher\Event\Theme\ThemesInstalledEvent $event */
     $event = $this->manager->getRegisteredEvent(HookEventDispatcherInterface::THEMES_INSTALLED);
     $this->assertSame($themes, $event->getThemeList());
   }
